Code generated from Sprint 1
The current version allows the user to get the PH of any liquid every 10 seconds or by requesting it manually. To get values immediately, press "r" and send it through the serial monitor. This version uses the millis function to determine when values should be gathered. After 49 days, the program will experience an integer overflow.